@@ -4,12 +4,10 @@ import android.graphics.Bitmap;
 import android.graphics.BitmapRegionDecoder;
 import android.graphics.Point;
 
-import com.koushikdutta.async.util.UntypedHashtable;
 import com.koushikdutta.ion.ResponseServedFrom;
 import com.koushikdutta.ion.gif.GifDecoder;
 
 import java.io.File;
-import java.nio.ByteBuffer;
 
 /**
  * Created by koush on 6/12/13.
@@ -28,12 +26,11 @@ public class BitmapInfo {
     final public String key;
     public ResponseServedFrom servedFrom;
     final public Bitmap bitmap;
-    public Exception exception;
+    public Throwable exception;
     public GifDecoder gifDecoder;
     public BitmapRegionDecoder decoder;
     public File decoderFile;
     public final String mimeType;
-    public final UntypedHashtable extras = new UntypedHashtable();
 
     public int sizeOf() {
         if (bitmap != null)
